Move island pruning to its own package
