+++ src/main/java/org/opentripplanner/graph_builder/module/islandpruning/Subgraph.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.graph_builder.module;
+package org.opentripplanner.graph_builder.module.islandpruning;
 
 import java.util.HashSet;
 import java.util.Iterator;
